feat: add executable (#3)
