+++ bin/sfcc-playground.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+const { run } = require('../src/app');
+
+run();
